chore: refactor code to add MinioUOW
